Renamed app and header
